update like functions

+++ src/modules/post/post.controllers.ts
@@ -83,16 +83,30 @@ const createPost = async (req: Request, res: Response) => {
 
 const getLikePost = async (req: Request<{ postId: string }>, res: Response) => {
 	const { postId } = req.params;
-
 	try {
+		const postExists = await prisma.post.findUnique({
+			where: {
+				id: Number(postId)
+			}
+		});
+		if (!postExists) {
+			return res.status(404).send({ message: 'Пост не найден' });
+		}
 		const likesCount = await prisma.like.count({
 			where: {
 				postId: Number(postId)
 			}
 		});
-		res.status(200).send({ postId, likesCount });
+		const userLike = await prisma.like.findFirst({
+			where: {
+				userId: req.user?.id,
+				postId: Number(postId)
+			}
+		});
+		const isLike = !!userLike;
+		res.status(200).send({ postId, likesCount, isLike });
 	} catch (error) {
-		console.error(error);
+		console.error('Ошибка при обработке запроса:', error);
 		res.status(500).send({ message: 'Внутренняя ошибка сервера' });
 	}
 };
@@ -100,6 +114,17 @@ const getLikePost = async (req: Request<{ postId: string }>, res: Response) => {
 const likePost = async (req: Request, res: Response) => {
 	const { postId } = req.body;
 	try {
+		const existingLike = await prisma.like.findFirst({
+			where: {
+				userId: req.user?.id,
+				postId: Number(postId)
+			}
+		});
+		if (existingLike) {
+			return res
+				.status(400)
+				.send({ message: 'Пользователь уже поставил лайк этому посту' });
+		}
 		const like = await prisma.like.create({
 			data: {
 				userId: req.user?.id!,
@@ -108,7 +133,8 @@ const likePost = async (req: Request, res: Response) => {
 				updatedAt: moment().utcOffset(6).format('YYYY-MM-DD HH:mm:ss Z')
 			}
 		});
-		res.status(200).send(like);
+		const { id, ...responseLike } = like;
+		res.status(200).send(responseLike);
 	} catch (error) {
 		console.error(error);
 		res.status(500).send({ message: 'Внутренняя ошибка сервера' });
@@ -118,6 +144,15 @@ const likePost = async (req: Request, res: Response) => {
 const unLikePost = async (req: Request, res: Response) => {
 	const { postId } = req.body;
 	try {
+		const existingLike = await prisma.like.findFirst({
+			where: {
+				userId: req.user?.id,
+				postId: Number(postId)
+			}
+		});
+		if (!existingLike) {
+			return res.status(404).send({ message: 'Лайк не найден' });
+		}
 		const like = await prisma.like.delete({
 			where: {
 				userId_postId: {
@@ -126,7 +161,8 @@ const unLikePost = async (req: Request, res: Response) => {
 				}
 			}
 		});
-		res.status(200).send(like);
+		const { id, createdAt, updatedAt, ...responseLike } = like;
+		res.status(200).send(responseLike);
 	} catch (error) {
 		console.error(error);
 		res.status(500).send({ message: 'Внутренняя ошибка сервера' });
